requirements.txt
altair
annotated-types
anthropic
anyio
attrs
blinker
cachetools
certifi
charset-normalize
click
colorama
distro
docstring_parser
gitdb
GitPython
h11
httpcore
httpx
idna
Jinja2
jiter
jsonschema
jsonschema-specifications
MarkupSafe
narwhals
numpy
openai
packaging
pandas
pillow
protobuf
pyarrow
pydantic
pydantic_core
pydeck
python-dateutil
pytz
referencing
requests
rpds-py
six
smmap
sniffio
streamli
tenacity
toml
tornado
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
watchdog
